+++ apps/react/demo/src/ui/Showcase/common/Example/hooks/useProviderState.ts
@@ -145,7 +145,7 @@ const useProviderState = ({
       copyOutput,
       activatePrevExample,
       activateNextExample,
-      demoOutput: demoOutput,
+      demoOutput,
       activeExampleFormValues,
       resetActiveExample,
       showBaselineGrid,