+++ packages/react/ds-app-launchpad/.storybook/preview.ts
@@ -4,6 +4,9 @@ import type { Preview, ReactRenderer } from "@storybook/react";
 import "index.css";
 
 const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: "^on.*" },
+  },
   decorators: [
     withThemeByClassName<ReactRenderer>({
       themes: {