+++ packages/react/ds-app-launchpad/src/ui/GitDiffViewer/common/CodeDiffViewer/CodeDiffViewer.stories.tsx
@@ -1,6 +1,7 @@
 /* @canonical/generator-canonical-ds 0.0.1 */
 
 import type { Meta, StoryFn } from "@storybook/react";
+import { fn } from "@storybook/test";
 import * as fixtures from "../../fixtures.js";
 import { GitDiffViewer } from "../../index.js";
 import type { ProviderOptions } from "../../types.js";
@@ -45,14 +46,24 @@ Default.parameters = storyOptions;
 export const WithComments = Template.bind({});
 WithComments.args = {
   diff: fixtures.diffExample,
-  lineDecorations: { 20: fixtures.commentExample },
+  lineDecorations: fixtures.commentExample,
 };
 WithComments.parameters = storyOptions;
 
 export const InteractiveGutterWithAddComment = Template.bind({});
 InteractiveGutterWithAddComment.args = {
   diff: fixtures.diffExample,
-  lineDecorations: { 20: fixtures.commentExample },
+  lineDecorations: fixtures.commentExample,
   AddComment: fixtures.addCommentExample,
 };
+
 InteractiveGutterWithAddComment.parameters = storyOptions;
+
+export const InteractiveGutterWithClickHandler = Template.bind({});
+InteractiveGutterWithClickHandler.args = {
+  diff: fixtures.diffExample,
+  lineDecorations: fixtures.commentExample,
+  onLineClick: fn(),
+};
+
+InteractiveGutterWithClickHandler.parameters = storyOptions;