+++ packages/react/ds-app-launchpad/src/ui/GitDiffViewer/common/CodeDiffViewer/types.ts
@@ -1,10 +1,16 @@
 /* @canonical/generator-canonical-ds 0.0.1 */
 import type React from "react";
 
-export type CodeDiffViewerAddComment = (props: {
-  lineNumber: number;
-  onClose: () => void;
-}) => React.ReactNode;
+export type CodeDiffViewerLineSelectOptions = {
+  hunkLineNumber: number;
+  diffLineNumber: number;
+};
+
+export type CodeDiffViewerAddComment = (
+  props: {
+    onClose: () => void;
+  } & CodeDiffViewerLineSelectOptions,
+) => React.ReactNode;
 
 export type CodeDiffViewerProps = {
   /** A unique identifier for the CodeDiffViewer */
@@ -12,9 +18,13 @@ export type CodeDiffViewerProps = {
   /** Additional CSS classes */
   className?: string;
   /** Add comment element.
-   * If provided, the add comment button will be displayed on the line number.
+   * If provided, the code viewer will be in interactive mode.
    */
   AddComment?: CodeDiffViewerAddComment;
+  /** Callback function for when a line is clicked.
+   * If provided, the code viewer will be in interactive mode.
+   */
+  onLineClick?: (options: CodeDiffViewerLineSelectOptions) => void;
   /** Inline styles */
   style?: React.CSSProperties;